+++ src/main/java/com/range/rpms/common/util/UserContext.java
@@ -3,4 +3,5 @@ package com.range.rpms.common.util;
 public interface UserContext {
 
     String getCurrentUserName();
+//    long getCurrentUserId();
 }